requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.5.0
apiclient==1.0.4
asgiref==3.7.2
astroid==2.6.6
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
black==23.7.0
blinker==1.4
cachetools==5.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
chromedriver-binary==105.0.5195.19.0
click==8.1.6
cloudpickle==2.2.1
contourpy==1.1.0
cryptography==37.0.2
currency.converter==0.5.5
cycler==0.11.0
Cython==3.0.0
distlib==0.3.2
dj-database-url==0.5.0
Django==4.2.3
django-debug-toolbar==3.2.4
django-heroku==0.3.1
django-mathfilters==1.0.0
django-nplusone==0.0.5
filelock==3.0.12
fonttools==4.41.0
frozenlist==1.4.0
fvcore==0.1.5.post20221221
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.0
grpcio==1.56.2
gspread==5.4.0
gunicorn==20.1.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
hydra-core==1.3.2
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
install==1.3.5
iopath==0.1.9
isort==5.9.3
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lazy-object-proxy==1.6.0
locate==1.1.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.6.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nplusone==1.0.0
numpy==1.22.4
oauth2client==4.1.3
oauthclient==1.0.3
oauthlib==3.2.0
omegaconf==2.3.0
openai==0.27.8
opencv-python==4.6.0.66
outcome==1.2.0
packaging==23.1
pandas==1.4.4
pathspec==0.11.1
pbr==5.6.0
Pillow==9.2.0
platformdirs==2.0.2
portalocker==2.7.0
protobuf==4.23.4
psycopg2==2.9.3
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.18.0
pylint==2.9.6
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
pyproject-toml==0.0.10
PySocks==1.7.1
pytchat==0.5.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.1
pyxel==1.8.8
PyYAML==5.1
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rpds-py==0.9.2
rsa==4.8
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.4
stevedore==3.3.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3
virtualenv==20.6.0
virtualenv-clone==0.5.5
virtualenvwrapper==4.8.4
webdriver-manager==3.5.4
Werkzeug==2.3.6
whitenoise==6.5.0
wrapt==1.12.1
wsproto==1.1.0
yacs==0.1.8
yarl==1.9.2
zipp==3.16.2
